added checks to the filemaking scripts

+++ makeday.py
@@ -3,10 +3,17 @@ import os
 from pathlib import Path
 import shutil
 
+allowedlanguages = ["python"]
+
+if (len(sys.argv) != 4):
+    sys.exit("ERROR: requires argv: year day language")
+
 year = sys.argv[1]
 day = sys.argv[2]
 language = sys.argv[3]
 
+if language not in allowedlanguages:
+    sys.exit("ERROR: allowed languages are: "+", ".join(allowedlanguages))
 
 cwd = os.getcwd()
 if not os.path.exists(cwd+"/"+year):
@@ -18,4 +25,7 @@ Path(cwd+"/"+year+"/"+day+"/"+day+".in").touch()
 Path(cwd+"/"+year+"/"+day+"/"+day+".test").touch()
 
 if language == "python":
-    shutil.copyfile(cwd+"/template.py",cwd+"/"+year+"/"+day+"/"+day+".py")
+    if not os.path.exists(cwd+"/"+year+"/"+day+"/"+day+".py"):
+        shutil.copyfile(cwd+"/template.py",cwd+"/"+year+"/"+day+"/"+day+".py")
+    else:
+        sys.exit("ERROR: "+cwd+"/"+year+"/"+day+"/"+day+".py already exists")
